Auto-migration in AvroDictManager, Plugin label strings updated to “Bangla Plugin”
