+++ .gitignore
@@ -102,3 +102,5 @@ lint/tmp/
 
 ### Kotlin ###
 .kotlin/
+*.md
+!README.md
