+++ app/src/main/java/org/fcitx/fcitx5/android/input/keyboard/bn/ProbhatKeyboard.kt
@@ -21,7 +21,7 @@ import org.fcitx.fcitx5.android.input.keyboard.LanguageKey
 import org.fcitx.fcitx5.android.input.keyboard.LayoutSwitchKey
 import org.fcitx.fcitx5.android.input.keyboard.ReturnKey
 import org.fcitx.fcitx5.android.input.keyboard.SpaceKey
-import org.fcitx.fcitx5.android.plugin.avro.AvroKeyboardIds
+import org.fcitx.fcitx5.android.plugin.bangla.AvroKeyboardIds
 import splitties.views.imageResource
 
 /**
