--- plugin/avro/org.fcitx.fcitx5.android.yml
@@ -1,4 +1,0 @@
-name: Avro Plugin
-components:
-  - name: plugin
-    type: plugin
