--- plugin/avro/src/main/cpp/avro-addon.conf.in.in
@@ -1,7 +1,0 @@
-[Addon]
-Name=Avro Bangla Phonetic
-Category=InputMethod
-Library=avro
-Type=SharedLibrary
-OnDemand=True
-Configurable=True
