--- plugin/avro/src/main/cpp/avro.h
@@ -1,55 +1,0 @@
-/*
- * SPDX-License-Identifier: LGPL-2.1-or-later
- */
-#ifndef FCITX_AVRO_AVRO_H
-#define FCITX_AVRO_AVRO_H
-
-#include <fcitx/addonfactory.h>
-#include <fcitx/addonmanager.h>
-#include <fcitx/inputcontext.h>
-#include <fcitx/inputmethodengine.h>
-#include <fcitx/instance.h>
-#include <fcitx/text.h>
-#include <fcitx-utils/misc.h>
-
-#include <string>
-#include <unordered_map>
-
-#include "avro_parser.h"
-
-namespace fcitx {
-
-class AvroEngine : public InputMethodEngine {
-public:
-    explicit AvroEngine(Instance *instance) { FCITX_UNUSED(instance); }
-
-    void keyEvent(const InputMethodEntry &entry, KeyEvent &keyEvent) override;
-    void activate(const InputMethodEntry &entry, InputContextEvent &event) override;
-    void reset(const InputMethodEntry &entry, InputContextEvent &event) override;
-
-private:
-    static bool isProbhatEntry(const InputMethodEntry &entry);
-    static bool isAvroInputChar(char ch);
-    void handleProbhatKeyEvent(KeyEvent &keyEvent);
-    void updatePreedit(InputContext *ic);
-    void clearBuffer(InputContext *ic);
-    void clearPreedit(InputContext *ic);
-    std::string resolveCommit() const;
-    void commitBuffer(InputContext *ic);
-    void loadUserDict();
-
-    AvroParser parser_;
-    std::string buffer_;
-    std::unordered_map<std::string, std::string> userDict_;
-};
-
-class AvroFactory : public AddonFactory {
-public:
-    AddonInstance *create(AddonManager *manager) override {
-        return new AvroEngine(manager->instance());
-    }
-};
-
-} // namespace fcitx
-
-#endif
