--- plugin/avro/src/main/cpp/probhat.conf.in
@@ -1,7 +1,0 @@
-[InputMethod]
-Name=Probhat
-Icon=fcitx-avro
-Label=Probhat
-LangCode=bn
-Addon=avro
-Configurable=True
