--- plugin/avro/src/main/res/values/strings.xml
@@ -1,6 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<resources>
-    <string name="app_name_release">Fcitx5 (Avro Plugin)</string>
-    <string name="app_name_debug">Fcitx5 (Avro Plugin | Debug)</string>
-    <string name="description">Avro (Bangla phonetic input method) engine support for Fcitx5</string>
-</resources>
