+++ plugin/bangla/org.fcitx.fcitx5.android.yml
@@ -1,0 +1,4 @@
+name: Bangla Plugin
+components:
+  - name: plugin
+    type: plugin
