+++ plugin/bangla/proguard-rules.pro
@@ -1,0 +1,1 @@
+-keep class org.fcitx.fcitx5.android.plugin.bangla.** { *; }
