+++ plugin/bangla/src/main/AndroidManifest.xml
