+++ plugin/bangla/src/main/cpp/bangla-addon.conf.in.in
@@ -1,0 +1,7 @@
+[Addon]
+Name=Bangla
+Category=InputMethod
+Library=bangla
+Type=SharedLibrary
+OnDemand=True
+Configurable=True
