+++ plugin/bangla/src/main/res/values/strings.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<resources>
+    <string name="app_name_release">Fcitx5 (Bangla Plugin)</string>
+    <string name="app_name_debug">Fcitx5 (Bangla Plugin | Debug)</string>
+    <string name="description">Bangla input methods (Avro phonetic and Probhat) for Fcitx5</string>
+</resources>
