+++ plugin/bangla/src/main/res/xml/plugin.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<plugin xmlns="../../../../../pluginSchema.xsd">
+    <apiVersion>0.1</apiVersion>
+    <domain>fcitx5-bangla</domain>
+    <description>@string/description</description>
+</plugin>
