+++ settings.gradle.kts
@@ -35,4 +35,4 @@ include(":plugin:chewing")
 include(":plugin:sayura")
 include(":plugin:jyutping")
 include(":plugin:thai")
-include(":plugin:avro")
+include(":plugin:bangla")
